OPTIMIZE autoloading of test Factories & Seeders
REFACTOR database test files to 'tests/database' directory.
