+++ tests/Assets/Providers/TestingServiceProvider.php
@@ -6,12 +6,17 @@ use Illuminate\Support\ServiceProvider;
 
 class TestingServiceProvider extends ServiceProvider
 {
+    /**
+     * Bootstrap any services.
+     *
+     * @return void
+     */
     public function boot()
     {
         // Publish migration file (if not already published)
         if (! class_exists('CreatePeopleTable')) {
             $this->publishes([
-                __DIR__.'/../migrations/create_people_table.php.stub' => database_path(
+                __DIR__ . '/../../database/migrations/create_people_table.php.stub' => database_path(
                     'migrations/'.date('Y_m_d_His', time()).'_create_people_table.php'
                 ),
             ], 'migration');
